Рефактор страницы регистрации
